requirements.txt
fastapi~=0.73.0
uvicorn~=0.17.0
pydantic~=1.9.0
SQLAlchemy~=1.4.29
asyncpg~=0.25.0
python-jose~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.5
aioredis~=2.0.1
aiohttp~=3.8.1
pydantic[dotenv]
pytest~=6.2.5
httpx~=0.22.0
pytest-asyncio~=0.17.2
ormar~=0.10.24
databases~=0.5.4
psycopg2~=2.9.3
asynctest~=0.13.0
mock~=4.0.3
flake8~=4.0.1
flake8-bugbear~=22.1.11
flake8-polyfill~=1.0.2
